ON-16509: mimic ef10 handling of missing time sync for ef(10)ct
With ef10 hardware, it's possible to request time-sync events without
the clock status (i.e., set/in-sync). This is also the case in ef10ct
(but not efct). To handle this appropriately, we only use the stored
clock status if we managed to request clock status updates (as ef10
does), and set this for efct whenever time-sync events are requested.