af_xdp: Perform getsockopt(SOL_XDP, XDP_MMAP_OFFSETS) during init_hardware
getsockopt writes to untrusted user memory, and if this is performed
during stack initialization in af_xdp_init by an adversarial user,
it's possible for them to race the memory being unmapped causing
potential OOB in the kernel.

On the other hand, af_xdp_nic_init_hardware may only be called by root.
The getsockopt may be performed earlier here. A temporary socket is
created to perform the getsockopt. Since the mmap offsets are just
struct offsets, the offsets will not change during runtime.

Signed-off-by: YiFei Zhu <zhuyifei@google.com>