ON-16532: determine evq time sync requirement per NIC
Previously, time sync events were required to get correct timestamps
for both RX and TX on ef10 hardware. As such, when disabling an evq,
we checked for having either of these enabled. However, this isn't
true for efct and ef10ct which only require time sync events for TX
timestamps. This patch creates a new EFHW function to determine if
a NIC requires time sync events based on EFHW flags. As a result, we
no longer invalidly try to disable time sync events for efct and
ef10ct if only RX timestamping is enabled.