ON-16481: fix data race in efct_vi_rx_future_peek
This function is called without a lock so could happen concurrently with
a kernel-side poll. This lead to a problem when reading the packet "pointer"
twice: once to check that it was valid (i.e. rollover not needed) and once
to access the next buffer. An update between these reads could lead to
accessing an invalid location beyond the end of the buffer.

Fix by reading the "pointer" just once and checking the same value for validity
that we will use for access.