ON-16630: Perform 16-bit read on efct rx metadata for filter_id
This read needs to be 16-bits as the filter_id is this large for x4.
Given that the user field in efct RX metadata is unused on x3, a 16-bit
read can be performed for the filter_id assuming these bits are zero. Do
not read user field from pkt metadata as it is unused.