ON-16659: Address X3 scheduling starvation in tcp_helper_wakeup
A problem this patch aims to address is a "soft deadlock" between
userspace holding a netif lock and the X3 interrupt handler in Onload,
which cannot acquire this lock and reprime interrupts synchronously,
triggering an interrupt on the same core and preventing userspace from
making any progress. To break this deadlock, as a workaround, it is
sufficient to change the affinity of either workflow.

This patch makes the interrupt handler reprime interrupts in the lock
contention case less eagerly, i.e. deferred in the workqueue. This, in
theory, should increase jitter in some unfortunate cases, but will
resolve the outlined deadlock scenario.

Reviewed-by: Josh Featherstone <Josh.Featherstone@amd.com>