ON-16677: Enabled LTO in release builds
Pass "-flto=auto -ffat-lto-objects" to the compiler and linker
if (GCC >= 11) and (NDEBUG is set).

This patch makes the standalone mmake userland build workflow
aligned with the RHEL10 rpmbuild workflow, which passes new flags
via CFLAGS.

Additionally, pass new flags in RHEL9, Ubuntu 22+, and Debian 12
"proactively", because it has potential for a performance gain in
the Onloaded applications.

The debug builds show slower runtime performance, and LTO nearly
doubles the build time. Hence, we enable it only for the release
builds.

The kernel build is unaffected, and most of the distro kernels
are built with CONFIG_LTO_NONE=y.