ON-16776: fix ulhelper build
The TCP RST rate limiting code is only used in tcp_rx.c, which itself
is only compiled in if OO_DO_STACK_POLL is set. This change wraps the
call to initialise ratelimiting with a check on OO_DO_STACK_POLL, as
well as the definition of the function itself. This fixes a build
error when building with ulhelper.