ON-16817: Set pd shared rxq token for onload stacks
Adds implementation to shrub_fns.c to get the token from the server and
update the pd with the token's value. Approach is very similar to
efct_ubufs_pre_attach, but modified to work in the kernel.

Additionally moves shrub_client.h from src/lib/ciul/ to
src/include/etherfabric/ to match shrub_server.h and so the functions
can be used in kernel space.