ON-16817: Connect to shrub server before binding onload stack to rxq
The first one to bind the rxq will determine which evq is used with the
rxq. Because of this the shrub server should be the first one to connect
so that an appropriate evq can be chose (especially because onload
stacks and likely to have a TXQ and so an EVQ which would be used).

Additionally, for subsequent rxq users get the status of the evq from
the existing state, rather than just relying on function parameters.