ON-15312: Add interrupt request flag to efct_rxq char driver interface
In preparation for full interrupt driven rx for shared rxq clients we'll
need to make sure that the rxq the clients attach to has interrupts set
appropriately. Previously, the interrupt_req parameter to efrm_rxq_alloc
was only able to be set in onload, but now the shrub controller (an
ef_vi app) will also need to set it when attaching to the rxq.

Subsequent patches will add the implementation details for shared rxq
wakeups on ef10ct.