ON-16548: Construct MCDI from stored filter info
We need to be able to translate any filter op to MCDI. This can be the result
of initial insert, redirect, or post-reset insert. The only information we
have consistently across those is the internal filter state. By constructing
MCDI from that state we only have to support doing conversion in one way.