ON-16868: refactor efct TX to use defined chunk size
Currently, efct transmits with an 8-byte write to the CTPIO aperture.
There are several assumptions throughout the transmit path for this.
This commit makes such assumptions clear by defining and using an
EFCT_TX_CHUNK_SIZE macro throughout, which will also make it easier
to change the transmit size in the future.