ON-16868: make ci_cpu_has_feature more generic
The current implementation of ci_cpu_has_feature matches a string and
then inspects the output of the required register to see which bit is
set. This notably only allows for checking of the 64 CPU features
included in the result of cpuid when EAX=1 (EDX and ECX are set).

This isn't particularly extensible as we may want to check other
features which require different input values for EAX (and ECX).
This commit refactors this code to allow a generic request with
options for setting the input values and specification of the
expected output register and bit.