ON-16868: make cpu_features header only
Currently the only consumer of the ci_cpu_has_feature function is in
the citools lib as well (toeplitz.c), but I want to use it elsewhere
in future commits. In particular, to check if movdir64b is supported
in efct_vi.c and the additional build-dependency on the citools lib
would be unacceptable. As such, I opted to move the code into the
header file as it was reasonably small.