ON-16868: fail efct vi creation in CXL profile if CPU lacks movdir64b
If the CPU does not support the movdir64b instruction, then it may
raise an illegal instruction signal (SIGILL) if it processes it.
This will definitely crash user apps, and might crash the kernel
(though I haven't verified this), so we should fail early and in
a way that informs the user of the issue.