ON-16548: Remove redundant filter functions
This gets rid of two things:
- the post add call for clusters, as this is not supported
- the replace parameter to post add

The replace parameter here is used to indicate to the net driver when we
install a filter it should be installed with a higher priority than any
existing equivalent filter. It doesn't make sense to pass it to post add,
where it is indeed unused.