ON-16159: rename EFRM_FLUSH_QUEUES_F_NOHW
I found it a little confusing to read without an additional underscore
between "NO" and "HW". This change also moves a comment relating to
software-only flushes to be alongisde this flag's definition instead
of above the efrm_nic_flush_all_queues function, which was appropriate
when it was made due to this being the only parameter.