ON-16159: store the txq number instead of ID in ef10ct
This patch both disambiguates the type of queue reference stored in
our evq structs and changes the type from an ID to a queue number.
In general, I prefer storing the queue number as it is more akin to
other architectures which don't clobber the top bits of the queue
ID to differentiate between the type of queue.