ON-13969: fix ef_vi_stats_query documentation
Since Onload 9.0 removed the ability to run on 7000 series of NICs the
note under ef_vi_stats_query() in the ef_vi documentation can never be
valid.

Previously for older NICs this function required full-feature
firmware. Since 8000 series it could run on ULL. However, the
documentation was not updated. Since the 7000 series is no longer
supported this note is now completely invalid. Therfore, this commit
removes it.