ON-16912: keep old efct tail location and alignment in non-CXL build
The flurry of CXL changes included moving the tail to the end of the
efct_tx_state struct and cacheline aligning the buffer. This change
led to a negative performance delta in some X3 use cases. To work
around this issue, place the tail back where it was before the CXL
changes and removes the alignment.