ON-16978: ON-16891: Separate out RXQ detach and local state RXQ cleanup
The cleanup RXQ op was inconsistent in whether it was cleaning up only local
client state, or shared vi state. This change makes it explicit that cleanup
is used only to cleanup local state. Where RXQs should be detached this should
be done separately, and before cleanup.

This fixes two issues:
1) use-after-free, where the local client state containing superbuf addresses
   was freed before we tried to unmap those buffers
2) hugepage leak on X3, because the fix to avoid detaching queues for all
   clients on X4 also meant we weren't cleaning up local state on X3. Now that
   cleanup only cleans up local state it is safe for use on both X4 and X3
   and so the previous commit required to fix X4 can be reverted.