ON-16808: ensure dispatchable syscall list remains up-to-date
Adds a build-time assertion that any use of DISPATCH_SYSCALLn is
accompanied by an entry in the FOR_EACH_DISPATCHABLE_SYSCALL
macro. This ensures the IBT validation in set_syscall_table is
unable to become stale without breaking the build, and without
having to perform additional checks at run-time, nor inclusion
of the __NR_syscalls macro which managed to evade me on several
kernel versions.