ON-16802: ignore double free of queues in efrm_vi_reinit_txq
It's possible that when calling efrm_vi_reinit_txq, we free our
software state but the subsequent queue initialisation fails,
in which case we likely want to try reiniting our queue again.
However, because the software state views the queue as freed,
we would never be able to try asking the NIC to initialise our
queue again, so we only update our state if necessary.