ON-17000: handle race in efct_vi_sync_rxq
When we first attach to a shared RXQ, we try to synchronise with all
other clients sharing this RXQ. This means that we want try updating
our state to be waiting for the next packet in the same place as the
other consumers. To do this, we find the first superbuf which is not
already full of packets, and then find the first packet which is not
used in that superbuf.

However, it is possible that by the time we start looking for the
first unused packet, we check every packet in the superbuf and
find that they have all been used. This isn't necessarily fatal,
but previously would result in a crash as we wouldn't try moving
on to the next superbuf resulting in freeing an efct packet we
though we had already freed.