ON-17016: disable TX error recovery on X3
When recovering from a TX error event on X3, we must tell the X3
driver that this event has happened to allow us to reinitialise
our TXQ - otherwise the driver believes that our TXQ has already
been initialised (which was true until firmware tore it down due
to the TX error event). This is done via efhw_nic_post_tx_error,
which is efct_nic_post_tx_error in the case of X3.

However, during the TXQ freeing process, the X3 net driver will
mark the EVQ that was being delivered TX events for this TXQ as
unused. This can result in various unfortunate mismanagement
issues, including filling the EVQ with 0xff bytes, which may
result in a kernel panic due to assumptions made by onload.