onload_remote_monitor: add logging of SYN_RECV state to orm_json output
There is no CI_TCP_SYN_RECV state in Onload. As said in Onload:

For each listening socket we have a list of SYNRECV buffs, one for each
SYN we've received for which there hasn't yet been an ACK.  i.e. on
receipt of SYN we make a synrecv buf, then send the SYNACK.  The on
receipt of the ACK of the SYNACK, we 'promote' the synrecv buffer to a
fully fledged end-point.  We call this list the listenq.

However information from this list was not logged in output of orm_json.
Therefore there was no information in orm_json output about sockets in
TCP_SYN_RECV state.

Signed-off-by: Nikolai Kosovskii <nikolai.kosovskii@arknetworks.am>