ON-17055: avoid spurious TXQ reinit retry attempt
If we fail to reinit our TXQ, and have run out of attempts, we no
longer need to queue the reinit work. Before this patch, we would
end up with one extra delayed work queued up after running out of
retry attempts, which was effectively a no-op, but it would be
better to avoid this.