ON-17056: keep reference to dshm after mmap
The donated shared memory mechanism allows an application to give the
onload kernel module access to a region of its memory. This can then
be mapped into the memory of another process, and read, via mmap.

During the memory mapping process, we incremented the reference count
of the shared memory region to avoid it being freed while we remapped
the pages into the new address space. However, we would then drop our
reference after finishing this. As a result, it was possible that the
application that donated the memory could exit and free this memory,
during which we would unpin these pages. As a result, an app could be
left with a stale mapping to a page that has already been unpinned.
On some newer kernels with debug options set, this could trigger an
assertion when the app with stale mappings was closed.

To avoid this, we maintain a list of shared memory regions a given
app has taken a reference to and drop them when the onload driver
handle is closed.