af_xdp: Suffix the XSP map pin paths
Each XSK and XDP prog are bound to a specific nic, so load the prog
multiple times, once for each NIC, and distinguish the XSKMAPs with
the ifname.

Updated API from bpf_object__pin_maps to bpf_map__pin in order to
specify the exact map to pin and the target file path; the former
API would use a filename matching the map name and would not suffix
the ifname.

Signed-off-by: YiFei Zhu <zhuyifei@google.com>