ON-17017: handle unfull efct_ubuf freeing
During normal operation, we go though ubufs in FIFO order. When a ubuf
becomes full, and the last packet is consumed, it will be freed by the
application. We may then repost that buffer to the NIC and move on to
the next buffer which the NIC will start filling up.

This flow is slightly modified when shutting down, as we immediately
become uninterested with all of our buffers. As a result, we iterate
through our buffers, freing each of them in turn with no regard for
whether they are full. This was an issue for efct_ubufs_free_local,
as it assumed we would only free a buffer if it was already full and
so should be handled in update_filled. When this wasn't the case, we
would end up putting the buffer on the free list and corrupting our
hardware fifo tracking state.

To avoid this problem, explicitly check for the case where we are
freeing a buffer that hasn't been handled by update_filled, and
update the necessary tracking state before adding the buffer to
the free list.